move typesVersions to types package
