+++ packages/amplify-v6-types-package/package.json
@@ -5,6 +5,13 @@
   "module": "./lib-esm/index.js",
   "types": "./lib-esm/index.d.ts",
   "license": "UNLICENSED",
+  "typesVersions": {
+		"<5": {
+			"lib-esm/index.d.ts": [
+				"index.v3.ts"
+			]
+		}
+	},
   "scripts": {
     "lint": "eslint \"**/*.ts*\"",
     "build:cjs": "rimraf lib && tsc -p ./tsconfig.build.json -m commonjs --outDir lib",
