+++ packages/api-graphql/package.json
@@ -56,7 +56,6 @@
     "@aws-amplify/core": "5.8.0",
     "@aws-amplify/pubsub": "5.5.0",
     "graphql": "15.8.0",
-    "tslib": "^1.8.0",
     "uuid": "^3.2.1",
     "zen-observable-ts": "0.8.19"
   },
@@ -65,7 +64,7 @@
       "name": "API (GraphQL client)",
       "path": "./lib-esm/index.js",
       "import": "{ Amplify, GraphQLAPI }",
-      "limit": "88.35 kB"
+      "limit": "90.46 kB"
     }
   ],
   "jest": {
