+++ packages/api-rest/package.json
@@ -48,7 +48,6 @@
   "dependencies": {
     "@aws-amplify/core": "5.8.0",
     "axios": "0.26.0",
-    "tslib": "^1.8.0",
     "url": "0.11.0"
   },
   "size-limit": [
@@ -56,7 +55,7 @@
       "name": "API (rest client)",
       "path": "./lib-esm/index.js",
       "import": "{ Amplify, RestAPI }",
-      "limit": "31.5 kB"
+      "limit": "31.99 kB"
     }
   ],
   "jest": {
