+++ packages/datastore/package.json
@@ -72,7 +72,7 @@
 			"name": "DataStore (top-level class)",
 			"path": "./lib-esm/index.js",
 			"import": "{ Amplify, DataStore }",
-			"limit": "137.1 kB"
+			"limit": "138.2 kB"
 		}
 	],
 	"jest": {
