+++ packages/notifications/package.json
@@ -66,13 +66,13 @@
 			"name": "Notifications (top-level class)",
 			"path": "./lib-esm/index.js",
 			"import": "{ Amplify, Notifications }",
-			"limit": "30.5 kB"
+			"limit": "30.6 kB"
 		},
 		{
 			"name": "Notifications (with Analytics)",
 			"path": "./lib-esm/index.js",
 			"import": "{ Amplify, Notifications, Analytics }",
-			"limit": "30.5 kB"
+			"limit": "30.61 kB"
 		}
 	]
 }
